+++ frontend/out/_next/static/chunks/app/layout-89411ba5c0f25751.js
@@ -1,0 +1,1 @@
+(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[177],{6204:()=>{},7497:(e,s,_)=>{Promise.resolve().then(_.t.bind(_,6204,23))}},e=>{var s=s=>e(e.s=s);e.O(0,[65,106,762,358],()=>s(7497)),_N_E=e.O()}]);