+++ frontend/out/_next/static/chunks/pages/_app-9bed07b30faff535.js
@@ -1,0 +1,1 @@
+(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[636],{2132:(_,n,p)=>{(window.__NEXT_P=window.__NEXT_P||[]).push(["/_app",function(){return p(5284)}])}},_=>{var n=n=>_(_.s=n);_.O(0,[593,792],()=>(n(2132),n(3666))),_N_E=_.O()}]);